fix: BrowerWindow before app ready
