+++ src/main/utils/assets-path.js
@@ -1,0 +1,8 @@
+import { app } from 'electron'
+
+const assetsPath =
+  process.env.NODE_ENV === 'production'
+    ? process.resourcesPath
+    : app.getAppPath()
+
+export { assetsPath }
